+++ dot/network/sync.go
@@ -1,10 +1,8 @@
 package network
 
 import (
-	"bufio"
 	"context"
 	"fmt"
-	"io"
 	"math/rand"
 	"reflect"
 	"sort"
@@ -421,57 +419,6 @@ func (q *syncQueue) receiveBlockResponse(stream libp2pnetwork.Stream) (*BlockRes
 	msg := new(BlockResponseMessage)
 	err = msg.Decode(q.buf[:n])
 	return msg, err
-}
-
-// readStream reads from the stream into the given buffer, returning the number of bytes read
-func readStream(stream libp2pnetwork.Stream, buf []byte) (int, error) {
-	r := bufio.NewReader(stream)
-
-	var (
-		tot int
-	)
-
-	length, err := readLEB128ToUint64(r)
-	if err == io.EOF {
-		return 0, err
-	} else if err != nil {
-		return 0, err // TODO: read bytes read from readLEB128ToUint64
-	}
-
-	if length == 0 {
-		return 0, err // TODO: read bytes read from readLEB128ToUint64
-	}
-
-	// TODO: check if length > len(buf), if so probably log.Crit
-	if length > maxBlockResponseSize {
-		logger.Warn("received message with size greater than maxBlockResponseSize, discarding", "length", length)
-		for {
-			_, err = r.Discard(int(maxBlockResponseSize))
-			if err != nil {
-				break
-			}
-		}
-		return 0, fmt.Errorf("message size greater than maximum: got %d", length)
-	}
-
-	tot = 0
-	for i := 0; i < maxReads; i++ {
-		n, err := r.Read(buf[tot:])
-		if err != nil {
-			return n + tot, err
-		}
-
-		tot += n
-		if tot == int(length) {
-			break
-		}
-	}
-
-	if tot != int(length) {
-		return tot, fmt.Errorf("failed to read entire message: expected %d bytes", length)
-	}
-
-	return tot, nil
 }
 
 func (q *syncQueue) processBlockResponses() {
