+++ dot/network/test_helpers.go
@@ -1,7 +1,7 @@
 package network
 
 import (
-	"bufio"
+	"io"
 	"math/big"
 
 	"github.com/ChainSafe/gossamer/dot/types"
@@ -122,44 +122,23 @@ func (s *testStreamHandler) handleMessage(peer peer.ID, msg Message) error {
 }
 
 func (s *testStreamHandler) readStream(stream libp2pnetwork.Stream, peer peer.ID, decoder messageDecoder, handler messageHandler) {
-	// create buffer stream for non-blocking read
-	r := bufio.NewReader(stream)
+	var (
+		maxMessageSize uint64 = maxBlockResponseSize // TODO: determine actual max message size
+		msgBytes              = make([]byte, maxMessageSize)
+	)
 
 	for {
-		length, err := readLEB128ToUint64(r)
-		if err != nil {
-			logger.Error("Failed to read LEB128 encoding", "error", err)
+		tot, err := readStream(stream, msgBytes)
+		if err == io.EOF {
+			continue
+		} else if err != nil {
+			logger.Debug("failed to read from stream", "protocol", stream.Protocol(), "error", err)
 			_ = stream.Close()
 			return
 		}
 
-		if length == 0 {
-			continue
-		}
-
-		msgBytes := make([]byte, length)
-		tot := uint64(0)
-		for i := 0; i < maxReads; i++ {
-			n, err := r.Read(msgBytes[tot:]) //nolint
-			if err != nil {
-				logger.Error("Failed to read message from stream", "error", err)
-				_ = stream.Close()
-				return
-			}
-
-			tot += uint64(n)
-			if tot == length {
-				break
-			}
-		}
-
-		if tot != length {
-			logger.Error("Failed to read entire message", "length", length, "read" /*n*/, tot)
-			continue
-		}
-
 		// decode message based on message type
-		msg, err := decoder(msgBytes, peer)
+		msg, err := decoder(msgBytes[:tot], peer)
 		if err != nil {
 			logger.Error("Failed to decode message from peer", "peer", peer, "err", err)
 			continue
